Refactor nft types
