+++ packages/@magic-sdk/types/src/index.ts
@@ -12,3 +12,4 @@ export * from './modules/auth-types';
 export * from './modules/rpc-provider-types';
 export * from './modules/user-types';
 export * from './modules/intermediary-types';
+export * from './modules/nft-types';
